refactor folder: move view components into "/ui/*"
