+++ code/system.ui/react.common/src/ui/Text.Input/TextInput.mts
@@ -1,0 +1,1 @@
+export * from './ui/TextInput';
