+++ code/system.ui/react.common/src/ui/Text.Input/index.ts
@@ -1,1 +1,2 @@
-export * from './TextInput';
+export type { TextInputProps } from './types.mjs';
+export * from './TextInput.mjs';
