+++ code/system.ui/react.common/src/ui/Text.Input/util.mts
